@@ -1,14 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "com.mkulesh.micromath"
         minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 38
-        versionName "1.20"
+        targetSdkVersion 29
+        versionCode 39
+        versionName "1.21"
         setProperty("archivesBaseName", "microMathematics-v" + versionName)
     }
 
@@ -27,8 +27,9 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-annotations:27.1.0'
-    implementation 'com.android.support:appcompat-v7:27.1.0'
-    implementation 'com.android.support:design:27.1.0'
+    implementation 'androidx.annotation:annotation:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.preference:preference:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation 'org.apache.commons:commons-math3:3.6.1'
 }
