@@ -18,8 +18,8 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -364,7 +364,7 @@ abstract public class BaseFragment extends Fragment implements OnClickListener
 
     public void updateModeTitle()
     {
-        android.support.v7.view.ActionMode mode = ((MainActivity) activity).getActionMode();
+        androidx.appcompat.view.ActionMode mode = ((MainActivity) activity).getActionMode();
         if (mode != null)
         {
             final int selected = formulas.getSelectedEquations().size();
