@@ -13,7 +13,7 @@
 package com.mkulesh.micromath.dialogs;
 
 import android.net.Uri;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.view.View.OnLongClickListener;
 import android.widget.CheckBox;
