@@ -13,7 +13,7 @@
 package com.mkulesh.micromath.formula;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageButton;
+import androidx.appcompat.widget.AppCompatImageButton;
 import android.util.AttributeSet;
 import android.view.ViewGroup;
 
